requirements.txt
uvicorn
fastapi
fastapi_utils
Flask < 2.0.0
Flask-Injector
PyJWT
PyYAML
Werkzeug < 2.0.0
alitra
azure-identity
azure-keyvault-secrets
azure-storage-blob
cryptography
dacite
flask-CORS
flask-restx
injector
numpy
python-dotenv
requests
transitions
